refactor: rename package utility to util
