+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/packets/play/SpawnEntityPacket.java
@@ -27,8 +27,8 @@ import xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion;
 import xyz.jonesdev.sonar.common.fallback.protocol.FallbackPacket;
 
 import static xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion.*;
-import static xyz.jonesdev.sonar.common.utility.ProtocolUtil.writeUUID;
-import static xyz.jonesdev.sonar.common.utility.ProtocolUtil.writeVarInt;
+import static xyz.jonesdev.sonar.common.util.ProtocolUtil.writeUUID;
+import static xyz.jonesdev.sonar.common.util.ProtocolUtil.writeVarInt;
 
 @Getter
 @ToString
