+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/packets/play/TeleportConfirmPacket.java
@@ -25,7 +25,7 @@ import lombok.ToString;
 import xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion;
 import xyz.jonesdev.sonar.common.fallback.protocol.FallbackPacket;
 
-import static xyz.jonesdev.sonar.common.utility.ProtocolUtil.readVarInt;
+import static xyz.jonesdev.sonar.common.util.ProtocolUtil.readVarInt;
 
 @Getter
 @ToString
