+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/packets/play/UniversalChatPacket.java
@@ -26,12 +26,12 @@ import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
 import xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion;
 import xyz.jonesdev.sonar.common.fallback.protocol.FallbackPacket;
-import xyz.jonesdev.sonar.common.utility.ComponentHolder;
+import xyz.jonesdev.sonar.common.util.ComponentHolder;
 
 import java.time.Instant;
 
 import static xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion.*;
-import static xyz.jonesdev.sonar.common.utility.ProtocolUtil.*;
+import static xyz.jonesdev.sonar.common.util.ProtocolUtil.*;
 
 @Data
 @NoArgsConstructor
